Work in progress: a script that generates GitMe-$VERSION.exe
Much like the installer I did for WinGit-$VERSION.exe, this script
makes a GitMe-$VERSION.exe.

The result will be slightly different from GitMe-3.exe: There is no
longer a reason to call sh.exe differently, and config.mak does not
have to be in a subdirectory.

Also, curl-config is only needed when compiling things (it shows you
the necessary CFLAGS and friends).

Furthermore, the generating script will be checked into msysgit.git,
so there is no need to include it into the installer.

This version is based on the second version of GitMe-3.exe...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>