GitMe: avoid some warnings, and initiate the "git" submodule better
There was a warning at the initial checkout, saying that we are on
a yet-to-be-born branch.  This startled some users.  So avoid it.
(This fixes issue 4.)

Also, do not try to be cute on the subproject and use read-tree.
Since there was no checked-out state yet, use the same checkout
command line as for the superproject.  (This fixes issue 45.)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>