WinGit: Add a native Win32 uninstaller
The easiest and cleanest solution for uninstalling WinGit would be
-- like the installer -- a Tcl script, since we already have Tcl
installed.

But that does not work, since Windows does not allow the removal of
files which are in use.

Therefore, we are _forced_ to have a standalone uninstaller, which
schedules its own removal at next reboot.

Incidentally, we now have an example how to write a non-bloated
graphical Win32 program in C.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>