WinGit: include html pages from official git.git's html branch
It is assumed that a matching version of the html documentation
is available as the HEAD of /doc/git/html/.git. If not an error
will be reported.

This patch doesn't include a mechanism to fetch the html pages
to /doc/git/html. The html pages should be at the right place,
which can be handled as a submodule of msysgit.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>