html help: add submodule html pointing to Junio's html for 1.5.3
We do not want to generate the html documentation in msysgit
but instead take the autogenerated documentation from Junio's
repo. This avoids having to integrate the complete toolchain
with msysgit.

So register a submodule at path doc/git/html, which points to
the html pages for 1.5.3 (which is our current git version in
msysgit).

To actually checkout the html pages you need to run

   git-submodule init
   git-submodule update doc/git/html

Signed-off-by: Steffen Prohaska <prohaska@zib.de>