WinGit(Inno): Make installer optionally modify PATH
By using a helper script, the installer can now optionally modify the current
user's PATH (which will be undone upon uninstall). Options are:

- Not to modify PATH at all,
- add only Git to PATH (i.e. \cmd),
- add Git and UNIX tools to PATH (i.e. \bin and \mingw\bin).

The default is to only add Git to PATH which seems to be the best compromise
between the convenience to be able to use Git from cmd.exe and avoiding
conflicts between programs with matching names (find, sort).