WinGit(Inno): Remove dependency on modpath.iss
The functionality to modify PATH was rewritten to better suit our needs (e.g.
to be able to first remove all occurrences of the installation directory from
PATH prior to installation). The code was simplified and merged into
install.iss. The drawback is that we lose Win9x support for now which
requires PATH to be set in autoexec.bat.

As briefly mentioned above, all paths that start with the installation path
are now removed from PATH before setting the new PATH as specified by the
user during setup. This avoids wrong PATH settings if a user upgrades or
reinstalls with different PATH settings that for the previous installation.

Moreover, the options if and how to modify PATH were now moved to a separate
wizard page during setup, along with more explaining comments.