WinGit(Inno): Set HOME if adding \cmd to PATH
If the user chooses to add \cmd to PATH in order to be able to use Git from
the Windows Command Prompt, HOME needs to be also set because /etc/profile
does not get executed in that case. This patch does this along with
introducing a little more generic methods to alter environment variables.

If HOME was set by the installer and it was not modified, the uninstaller will
remove it again, just like it reverts changes to PATH.