WinGit(Inno): Set HOME only if adding \bin to PATH (not for \cmd)
It is sufficient to set HOME if the user chooses to add \bin to PATH
in order to be able to use Unix tools from the Windows Command Prompt.
Setting HOME is not needed if only \cmd is added because the wrappers
in \cmd set HOME before they execute the wrapped command.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>