WinGit(Inno): Quit uninstaller when ssh-agent is running
If ssh-agent is running in the background, uninstall would not be able to
remove it, leaving files and directories on disk after uninstall. This is
fixed by issuing a warning before uninstall starts if ssh-agent.exe cannot
be deleted, asking the user to stop all ssh-agent processes and run
uninstall again.

[sp: adjusted ReleaseNotes.txt ]

Signed-off-by: Sebastian Schuberth <sschuberth@visageimaging.com>
Signed-off-by: Steffen Prohaska <prohaska@zib.de>