WinGit(Inno): Delete HOME if a previous installation modified it
If an update was performed (i.e. running the installer a second time without
uninstalling in between) with different options, it was the case that HOME
might be set although not required be the updated install. This is solved by
deleting HOME if a previous installation already modified it, before setting
it possibly again, similar to what is done for PATH.

[sp: adjusted ReleaseNotes.txt]

Signed-off-by: Sebastian Schuberth <sschuberth@visageimaging.com>
Signed-off-by: Steffen Prohaska <prohaska@zib.de>