Add SVG versions of the MSys logo, and the newly created msysGit logo
To make the difference between (Win-)Git and msysGit clearer, I use
a different logo for the latter, which is constructed by overlaying
the MSys "M" with the Git logo.

For future proofing, these logos were designed in Inkscape, and saved
as SVG files.

Reading this SVG with Gimp, rendering it at 32x32 and saving it as
an ICO file yields a logo that hopefully is liked by other people, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>