WinGit(Inno): Lower user rights required to install
Depending on the directory to install to, the user might require admin or
poweruser rights to install. But for installing to directories owned by the
user, no special rights are necessary. Installing into those directories was
not always possible before. This is hopefully fixed by explicitly setting the
required privileges to "none" and only writing to user specific locations.