WinGit(Inno): Installing as admin creates icons for all users
Previously, the changes made by the installer to the system always only
affected the user running the installer, even if that user was the admin. If
an admin is now running the installer, all icons and registry changes (this
includes the environment variables and shell extensions) are now performed for
all users. The only excpetion is the Quick Launch icon, as there is no common
users conceopt for that. For non-admins nothing has changed.