Add /share/tcltk/release.sh, a script to update Tcl/Tk
Start with a clean working directory.  Really clean (no untracked
files).  Then change the script to contain the desired version.
Commit this.

Then run it; since the state is pristine, it can deduce automatically
which files were installed, adding them to the index and overwriting
/share/tcltk/fileList.txt.

It is your honor to commit the result (you can amend the update to
release.sh if you feel like it).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>