Rename installperl.sh into release.sh and add install stuff
This is mainly ripped from the Tcl/Tk work I did, to bring the same
functionality to Perl.

Basically, edit /share/perl/release.sh to reflect the Perl version you
want to install, then commit it (this is necessary, since this script
does not even allow untracked files in the working tree).

Then run it, and it will fetch/build/install Perl, and even update the
index (as well as /share/perl/fileList.txt which was used to uninstall
the previous version).  You can then commit this as-is.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>