WinGit(Inno): Fix launching "Git Bash" for Win XP 64
Under Windows XP x64 Edition, the default Command Prompt is a 64-bit process,
whereas MinGW's sh.exe needs to be launched as a 32-bit process. This patch
explicitly calls the 32-bit cmd.exe from the SysWOW64 under a 64-bit OS in
order to launch sh.exe to fix this. Note that under Windows Vista 64-bit this
patch would not be needed as the OS seems to automagically find out that
sh.exe needs to be called as a 32-bit process.