WinGit: Teach setup to display release notes in separate window
Setup used to display the release notes in a text box
inside the setup application.  This text box could not
be resized, which made the release notes hard to read.

This commit modifies the setup to lauch WordPad for
displaying the release notes.  The user can resize
the window.  She can also copy text.  Overall, the
release notes are more useful than before.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>