WinGit: Delete wrapper scripts as they might have been replaced by built-ins
Executing a script takes precedence over executing a .exe file in MSYS when
omitting the extension. If a newer Git version promoted a script to a built-in,
that version will no longer come with the script version of the command. Thus,
when performing an update instead of a new install, the script from the older
version will be present as well as the new git.exe with the command built-in.
However, typing in the command's name at the shell will still launch the script
instead of (a copy of / a link to) git.exe. Therefore we now delete all scripts
that have the name of a built-in before we create the built-in copies / links.