Append PATH in git.cmd and gitk.cmd
The old way of handling PATH in git.cmd and gitk.cmd was not perfect.
git.cmd and gitk.cmd used to completely isolate the msysgit PATH from
the system PATH.  One problem is that programs outside of the msysgit
installation must be called with their full path, even if they are in
the system PATH.

This commit appends the system PATH after the msysgit paths, as
suggested by Clifford Caoile.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>
Tested-by: Clifford Caoile <piyo@users.sourceforge.net>