Tweak msys_p2w to better handle strings that resemble path lists
This commit modifies msys-1.0.dll's path conversion to avoid
mangling of paths containing ":", for example
"c:/path/to/repository".  Due to the ":", such a path looks like a
path list, but the path should not be modified.  The two patches
part of this commit contain the changes to msys_p2w().