Fix msys_p2w() tweak for path lists
This commit fixes the path list heuristic modified by the previous
commit.  The previous commit modified the heuristic to leave all
strings that do not start with "." nor "/" as is.  But this is wrong:
Those strings must not be considered path lists, but they might need
to be modified nonetheless, for example if they contain arguments with
assignments, like git's "--template=/path/to/tmpl".