Add script that semi-automatically builds a patched msys-dll from source
The script that builds msys-dll is named release.sh.  It downloads
MSYS-1.0.11-20071204-src from sourceforge; unpacks the source and
commits it to a freshly created git repository; applies our patches;
and outputs instructions how to build the patched source.

The source release of MSYS has a slightly different directory layout
than MSYS's CVS, so this commit also adjusts the paths in our patches.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>