Switch to new 'libexec/git-core' directory layout
Git 1.6.0 will no longer place the dashed forms in 'bin' but
in 'libexec/git-core'.

This commit adjusts msysgit to support the new directory layout.
It also updates the git subproject to the first commit that
installs programs in the new location.  Note that this commit
only makes the basic adjustments.  A separate commit will modify
the installer.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>