HowToRelease: Revert to making releases from master
This commit reverts e98ea7df1e53d3bdbe2282450b2db10a7d76962b.
Creating releases from a branch that is regularly rebased is confusing
and was not used in practice.  Dscho continued to create releases
using the old process that is based on two branches master and devel,
which is now documented in the howto.