Add helpers to install MinGW packages
The pre-install.sh and post-install.sh helper are meant to help the stuff
in /src/, of course.

Call pre-install.sh before installing stuff.  This script will output the
name of a temporary Git index which contains the current state.

Now install whatever you want.

Afterwards, call post-install.sh with the name of the temporary Git index
and a commit message as arguments; this will commit the changes.

The advantage to the earlier installation method of packages in /src/ is
that you can do that with a non-pristine state (in other words, you can
have untracked files lying around; they may be overwritten, though).

Note: the use of the plumbing 'diff-files' in post-install.sh is
crucial; we do _not_ want the index to be auto-refreshed; instead,
we want to catch all files that have been touched in the meantime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>