Fall back to / if the directory $USERPROFILE cannot be accessed
If the user has non-ASCII characters in her name, the variable $HOME
(being set to $USERPROFILE) will disagree with the encoding of the
actual path as it is on-disk.

The reason is that the environment variables adher to the current locale
(e.g. CP1252), but the file system stores file names in UTF-16.

As long as MSys is not modified to handle this case as the user would
expect, we fall back to setting HOME=/ (which is better than nothing).

This works around msysGit issue 108.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>