/etc/profile: set default protocol for plink to ssh
Plink can be configured to default to the telnet protocol, but of course,
there will never be a Git daemon listening on the telnet daemon.  It has
been reported on the msysGit issue tracker that setting the environment
variable PLINK_PROTOCOL to 'ssh' would fix the issue.

Since it cannot hurt to set that environment variable, let's do it and
hope that it works as expected.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>