share/WinGit/release.sh: provide a few safeguards
There are some pitfalls when releasing a new installer: as we want to
be able to recreate every installer, the version from which the
currently installed Git was built needs to be tagged.

Check that (and a few other things), and require the flag "--force" if
any precondition is not met.

The --force flag is needed so that it is still easy to make a quick
"release", e.g. when a bug is fixed (so the bug reporter can verify the
fix).

The order of the checks was chosen by the following measure: the more
interesting the check could be even if the user chose to pass --force
later, the earlier the check is performed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>