Add a script that checks if commits maintained literal copies
When updating, say, gcc, it is possible that some files were updated, but
others were left as-are.  This is particularly wrong with files that are
supposed to be hardlinks (but that are actually copies due to being
checked out by Git).

Provide a script that checks if commits left literal copies intact.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>