Fix resetting the origin of msysgit repo in net installer
After the first successful fetch of MSys environment over git:// the
origin was reconfigured to use http:// causing further fetches to be
extremely slow and CPU-intensive.

The issue was caused by incorrect order of applying || and && operators
when trying various options to fetch the environment. Essentially, when
the first git fetch succeeded, USE_HTTP was not set (so MinGW Git sources
were configured to use git://), but the origin was nevertheless changed
and the second git fetch was executed. But the second git fetch does not
produce any output and exits successfully, kind of hiding the issue until
the [much] later fetch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>