Add a patch to msys-1.0.dll to ignore CYGWIN=tty
There is no working pty emulation in MSys, so it makes no sense to
allow asking for that emulation.  The symptoms are non-working vi,
looping ssh questions to accept host keys or passwords.

Typically, CYGWIN=tty is set on boxes where CYGWIN is installed, so
we should not require the user to unset it manually.

This fixes msysGit issue 138, and possibly 96, 192, 227, 243, 246,
248 and 251.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>