/etc/profile: use $HOMEDRIVE\$HOMEPATH as default HOME
Heiko Voigt changed msys-1.0.dll to leave HOMEDRIVE and HOMEPATH alone, so
we should actually use it, too.

While at it, use this default not only if the variable HOME is unset, but
also if it does not point to a directory.

This should finally take care of msysGit issue 108.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>