Properly fix the 'cd $HOME' issue
When starting the Git Bash or Git GUI from the start menu or a quick
launch icon, we would like it to change the current working directory
to the user home, but at the same time, the shell should _not_ change
the current working directory when called from somewhere else.

Unfortunately, we cannot call bash without --login, and there is simply
no way to pass an option via the command line with that option, and
we cannot modify the environment in .lnk files.

But what we actually want to do is to set the current working directory,
which _can_ be done from .lnk files, even to paths specified by
environment variables.

So let's just set the current working directory to %HOMEDRIVE%%HOMEPATH%.

This is a slight change in behavior: if the user set the environment
variable HOME to something different, the initial current working
directory will not be what HOME says.  We'll just have to live with that.

Thanks go to Sebastian Schuberth for help with the InnoSetup installer
and Heiko Voigt for pushing the issue forward.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>