Add a helper to compile & install OpenSSL (MSys)
This is an adaption of the corresponding MinGW script, but we can use
the standard (Configure) way to set things up.

Configuring with --prefix= does not work (as it uses the default, /local),
and neither --prefix=/ (as it tries to "mkdir //ssl", the double slash
being misinterpreted by MSys).  So use --prefix=/usr instead, because
/usr is mounted to / anyway.

This installs the .dll files with a "msys-" prefix, as mandated by MSys.
That means that it overwrites the msys-*.dll files needed by ssh.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>