Git installer: Only offer to use Plink if at least one PuTTY session is found
This is to prevent Windows users to select Plink just because it sounds more
familiar, although they do not really use PuTTY. For users not using PuTTY,
Plink is more tricky to set up correctly which is a cause of support requests.
Those users should just go with OpenSSH which works out of the box.