Fix the window icon and title for "Git Bash" when launched via shell extension
Windows sets the initial icon and title for command line windows to the icon
and name of the shortcut that launched the program. So, in order to get a
proper icon and title for "Git Bash Here", it needs to launch a shortcut. As
the Regitry's "command" key cannot launch shortcuts directly, we need a wrapper
in the form of a Visual Basic Script that uses the ShellExecute function to do
so. That script also changes the current directory as needed and launches a
special version of the "Git Bash" shortcut that does not set a working
directory, so the current directory is not overridden.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>