install git-cheetah explorer extension by default
git cheetah provides more options to interrogate the history of files
with git than the simple registry context menu items. Especially if you
only want to know about part of the repository (e.g. one file)

The registration of git-cheetah needed to be implemented using the
script engine of InnoSetup because Windows does not allow to delete
files that are in use by running processes (the explorer in this case).

The native support of InnoSetup for actively used files uses a Windows
mechanism that registers a file for replacement during restart. The
drawback is that this mechanism needs Admin Privileges.

To circumvent this situation we rename the old dll before installation.
This way we can install and register the new dll without the need for a
complete restart. Simply restarting the explorer is sufficient and the
easiest method to do that is to logout and back in again.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>