Installer: Fix the CreateHardLink signature for Unicode builds
We need to call CreateHardLinkW for Inno Setup Unicode, where the "String"
data type is a Unicode string.

Moreover, load CreateHardLink[AW] delayed, and only during setup. Note that
in this particular case "delayload" is not necessary, because the call
to CreateHardLink is guarded by "IsNTFS", so it should always be available.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>