Installer: Use Vista's Symbolic Links for built-ins, if possible
Unfortunately, Windows Explorer does not report the correct directory size if
hardlinks are used inside that directory, which caused user complaints in the
past, stating an installed msysGit would be as large as 150 MB (whereas it is
more about 60 MB). This is why we prefer to use Symbolic Links now, if they
are available, as those are handled correctly by Windows Explorer.

Other advantages of Symblic Links are that they can be relative, and point to
a different volume. Disadvantages are that they become invalid if the target is
deleted, they cannot be easily enumerated (a target does not know its links),
and creating them requires administrator rights / user elevation.

If built-ins cannot be created using Symbolic Links for some reason, Hardlinks
are tried next, and if that also fails, a copy is created.

Credits go to Michael Lukashov for the suggestion and initial patch.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>