/src/rt/release.sh: make sure that the correct cc.exe is used
Even directly after a switch to the 'msys' branch, the compilation
should not fail miraculously. But it did.

The reason was that the PATH was not reset just by switching branches,
even if starting a new bash would set PATH correctly in /etc/profile.

So, not only set the MSYSTEM, but also the PATH appropriately in our
msys-1.0.dll build script, and everybody's happy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>