Add a script to help the transition MSysGitWiki -> GitWiki
We want to move the Wiki pages to the GitWiki, as the latter has a nice
and powerful MediaWiki engine, and it is also much easier to get people
involved.

This script downloads the old pages, transforms them to MediaWiki markup
(the transformation is ad-hoc, just enough for what this developer tested),
opens the result in a graphical editor for easy Ctrl-A Ctrl-C (select all
& copy), and also opens the appropriate link in the web browser, where
the copied text can be simply pasted, previewed and saved.

Once the editor is closed, the next page is handled.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>