Propagate error codes from the git executable.
Any error code from running git was masked by the chcp call to restore the
codepage. This propagates the code from the git program by explicitly
setting the ERRORLEVEL environment variable. This then overrides the
ERRORLEVEL value that might be set by chcp and is eventually emitted by
the final command. ERRORLEVEL and %ERRORLEVEL% are not the same thing.
See http://blogs.msdn.com/oldnewthing/archive/2008/09/26/8965755.aspx

Also quote the COMSPEC to handle spaces in the expansion path and make it
possible to pass more than 8 args to git-gui.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>