Add a script to perform a 'rebasing merge' in /git/
A 'rebasing merge' is a rebased branch, but merging the original state
via the 'ours' strategy so that checkouts can be fast-forwarded easily.

Of course, a 'rebasing merge' should avoid to rebase commits incurred
by a previous 'rebasing merge'.

As our rebase target is junio/next, which itself is a frequently rebased
branch, we have special handling to exclude commits committed by known
non-msysGiticists.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>