add-shortcut.tcl: Determine the MSYS root directory without using $INPUTRC
The add-shortcut script has been finding the MSYS root directory based
on the value of the INPUTRC environment variable, based on the
assumption that $INPUTRC is always set to "<root>/etc/inputrc".  But
this may not be the case.  If "$HOME/.inputrc" exists, /etc/profile
will preferentially point $INPUTRC at it.  That will cause the
shortcuts to be generated incorrectly.

There are a few different ways to find the MSYS root; this simply
assumes the root is at ../.. from the script directory.

Signed-off-by: bert Dvornik <dvornik+git@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>