Add script to compile/install/distribute antiword
This days Git on Windows is used not only for code, but more and more
for managing documentation workflow. And on Windows a lot of people use
*.doc as document format.

So the idea is to provide textconv diff & blame support for *.doc out of
the box, for which we'll need default textconv driver.

Personally I've been using antiword on Linux for such porpose for quite
some time already, and while not perfect, antiword works ok -- I can see
what's changed in git, and `git gui blame` docs, which is great to me.

NOTE: antiword needs to access files with encoding mapping tables, and since we
are going to add antiword to distribution (next patch) for this I had to patch
it to determine installation prefix at runtime, just like git does.

Signed-off-by: Kirill Smelkov <kirr@mns.spb.ru>