Add script to compile/install poppler (for pdftotext)
Like with *.doc, *.pdf is a widely used format, so it would be a good
idea to ship default textconv driver for pdf. For this we'll use
pdftotext which is part of poppler. pdfinfo is also sometimes useful, so
let's ship it too.

NOTE: libpoppler-7.dll is 4.7M unstripped and 1.6M stripped. So far
I've not stripped it, becase right now we don't strip other
libraries and programs which we build ourselves from src/ (e.g.
libcurl-4.dll is 403K as shipped and 297K if stripped). But maybe it
would be a good idea to strip everything we build...

NOTE2: I've had to workaround one bug in TDM GCC that we ship...

Signed-off-by: Kirill Smelkov <kirr@mns.spb.ru>