Associate .sh files to be run with bash in the installer
This adds a checkbox to the installer to associate *.sh files with
/bin/sh.exe.

Having bash scripts automatically run by git bash is helpful because you
can directly bash scripts from cmd.exe, via doubleclick in the explorer
or use them directly in IDEs.

We add a DropHandler for the COM object of shell32.dll which will
dispatch the dropped files as arguments to the registered open command
of .sh files. Thus you can drop files on sh scripts the same ways as on
batch scripts.

The registered icon is the default icon for batch scripts to give some
visual indication that .sh files are something similar and can be
executed directly.

Signed-off-by: Heiko Voigt <heiko.voigt@mahr.de>
Cleaned up and
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>