Add external manifests for all executables
On Windows Vista, not having an application manifest with a
requestedExecutionLevel can cause several kinds of confusing behavior.

I sent a separate patch to add an embedded manifest to the git
executables. This patch adds a similar external manifest to the rest of
the executables. It detects whether the executable already contains an
embedded manifest before creating the manifest file. The ideal situation
would be for every executable to have an embedded manifest, but
meanwhile this can avoid having to change every build system.

This patch alone will not work as well for the git executables, because
some of them are created by the installer (by copying or hardlinking).
If used together with the previous patch, it should cover every
executable.

Tested by running copy-files.sh by hand on Wine.

Signed-off-by: Cesar Eduardo Barros <cesarb@cesarb.net>