Perform a more elegant check for msys-1.0.dll's MD5
Revert the change to copy-files.sh done in 678125b that was meant to fix the
MD5 comparison done for msys-1.0.dll in /etc/profile, which was broken due
to the way the file path is written since 42a6098.

Instead, do not do a plain text comparison of MD5 sums, but let md5sum
resolve the file path and do the comparison itself.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>