rebasing-merge: specify all commits explicitly
The earlier attempt to specify additional commits included way too many
commits. So let's start to specify all of the commits explicitly.

While at it, refactor a little and offer a --graph mode that does not
perform the rebasing merge, but shows the commits to be rebased intead.

In the end, it looks like a complete rewrite of the logic, though.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>