etc/profile: set MAGIC only if the magic file exists
commit 006d11d (Help 'file' find its magic file) added the MAGIC
variable to /etc/profile. That makes sense in the msysgit development
envronment but not for the installed version (because 'file' doesn't
get installed).
This checks for the existence of the magic file before setting MAGIC.
That way it's possible to use another 'file', e.g. from Cygwin inside
the git bash.

Signed-off-by: Stefan Naewe <stefan.naewe@gmail.com>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>