run-tests: remove MSVC-leftovers which would make the tests fail
When launching "/path/to/git" and there exists both a directory "git" as
well as an executable "git.exe", under Windows the directory is preferred.
Since the MSVC stuff is written into directories of the basename of the
generated executable, we have to remove these directories before running
the tests.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>