msysGit: Use a more generic way to change to the working tree root
When not using the msys.bat environment but another git installation to do
a rebase, the absolute "cd /" would not change to the working tree root
and cause git diff to fail with a usage information like

    usage: git diff [--no-index] <path> <path>

So use a more generic way to determine the working tree root within the
hook code that will work with another git installations, too. Note that
"git rev-parse --show-cdup" will return an empty string (not ".") if the
current directory is the working tree root, so we need to prefix the
resulting string with "./" in order for "cd" to work as expected.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>