libiconv: Use the new-style signature for iconv() to match upstream MinGW
If mingw32-libiconv is installed via minigw-get, iconv() as declared in
iconv.h uses the new-style non-const inbuf as second argument. This is
achieved by exporting some automake variables in libiconv-*.mgwport. Do it
the same way here, too, so that we can have OLD_ICONV turned off when
compiling Git either in msysGit or mingwGitDevEnv.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>